Rename lynxkite_graph_analytics.ops to operations. This fixes the docs build error.
