+++ docs/reference/lynxkite-graph-analytics/operations.md
@@ -1,8 +1,8 @@
-::: lynxkite_graph_analytics.ops.basic_ops
-::: lynxkite_graph_analytics.ops.file_ops
-::: lynxkite_graph_analytics.ops.graph_ops
-::: lynxkite_graph_analytics.ops.ml_ops
-::: lynxkite_graph_analytics.ops.pykeen_ops
-::: lynxkite_graph_analytics.ops.query_ops
-::: lynxkite_graph_analytics.ops.visualization_ops
+::: lynxkite_graph_analytics.operations.basic_ops
+::: lynxkite_graph_analytics.operations.file_ops
+::: lynxkite_graph_analytics.operations.graph_ops
+::: lynxkite_graph_analytics.operations.ml_ops
+::: lynxkite_graph_analytics.operations.pykeen_ops
+::: lynxkite_graph_analytics.operations.query_ops
+::: lynxkite_graph_analytics.operations.visualization_ops
 ::: lynxkite_graph_analytics.networkx_ops
